+++ pushno/exceptions.py
@@ -8,3 +8,9 @@ class InvalidParamValue(Exception):
     """
     invalid parameter value
     """
+
+
+class InvalidService(Exception):
+    """
+    invalid service
+    """