+++ pushno/plugins/baseplugin.py
@@ -4,11 +4,13 @@ import requests
 
 
 class AbstractClient(ABC):
-    def __init__(self, api_key="", user_key=""):
-        self._api_key = api_key
-        self._user_key = user_key
+    def __init__(self, **kwargs):
         self._s = requests.Session()
 
     @abstractmethod
     def send(self, message):
+        pass
+
+    @abstractmethod
+    def validate_user(self, **kwargs):
         pass