+++ CHANGELOG.rst
@@ -5,3 +5,10 @@ Version 0.0.1 (2019-04-17)
 --------------------------
 
 * initial relase with support for PushOver and prowl
+
+
+Version 0.0.2 (2019-04-18)
+--------------------------
+
+* added common PushNotification class to access all available
+  services in a consistent fashion